requirements.txt
fastapi
uvicorn
python-multipart
openai
neo4j
PyMuPDF
Pillow
pydantic
python-dotenv
aiofiles
websockets